[FEAT]: вынесла константы и типы отдельно
